refactor: moved tests and update data attr for registration events
